pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1"
description = "My backend project"
authors = ["dx0naclyoo <konyzhenkovdanil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.26"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pydantic-settings = "^2.1.0"
jinja2 = "^3.1.3"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
